style: extract types to a separate file
